pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spellbook-py"
version = "0.0.1"
authors = [{ name = "Peter Podsky", email = "piotr@holobits.pl" }]
maintainers = [{ name = "Peter Podsky", email = "piotr@holobits.pl" }]
description = "Personal set of python tools used for varying purposes, like arranging my own Twitter bookmarks, developing AI tools for myself etc."
readme = "README.md"
keywords = ["python", "AI", "personal", "learning"]
requires-python = ">=3.9"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: MacOS",
  "Operating System :: POSIX :: Linux",
  "License :: OSI Approved :: Unlicense",
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
]
dependencies = [
  "python-dotenv >= 1.0.0",
  "requests >=2.31.0",
  "requests_oauthlib >= 1.3.1",
]

# [project.optional-dependencies]
# visualization = ["manim >= 0.17.2"]
# test = ["pytest-cov ~=3.0.0"]

#[project.scripts]
#my-script = "my_package.module:function"

# [project.urls]
# "Homepage" = "https://github.com/cft-nisq/qrem"
# "Bug Tracker" = "https://github.com/cft-nisq/qrem/issues"
# "Documentation" = "https://github.com/cft-nisq/qrem/docs"
# "Tutorials" = "https://github.com/cft-nisq/qrem/tutorials"

[tool.setuptools.packages.find]
where = ["src"] # list of folders that contain the packages (["."] by default)
